call method by name inside a class plus variables types changed

+++ src/test/java/ru/netology/stats/SalesCalculateTest.java
@@ -9,8 +9,8 @@ public class SalesCalculateTest {
     void minMonthTest() {
         SalesCalculate service = new SalesCalculate();
         long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
-        long expected = 9;
-        long actual = service.minSales(sales);
+        int expected = 9;
+        int actual = service.minSales(sales);
         Assertions.assertEquals(expected, actual);
     }
 
@@ -18,8 +18,8 @@ public class SalesCalculateTest {
     void maxMonthTest() {
         SalesCalculate service = new SalesCalculate();
         long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
-        long expected = 8;
-        long actual = service.maxSales(sales);
+        int expected = 8;
+        int actual = service.maxSales(sales);
         Assertions.assertEquals(expected, actual);
     }
 
@@ -27,8 +27,8 @@ public class SalesCalculateTest {
     void summaryTest() {
         SalesCalculate service = new SalesCalculate();
         long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
-        long expected = 180;
-        long actual = service.summary(sales);
+        int expected = 180;
+        int actual = service.summary(sales);
         Assertions.assertEquals(expected, actual);
     }
 
@@ -36,8 +36,8 @@ public class SalesCalculateTest {
     void averageSalesTest() {
         SalesCalculate service = new SalesCalculate();
         long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
-        long expected = 15;
-        long actual = service.averageSalesValue(sales);
+        int expected = 15;
+        int actual = service.averageSalesValue(sales);
         Assertions.assertEquals(expected, actual);
     }
 
@@ -45,8 +45,8 @@ public class SalesCalculateTest {
     void monthsSalesOverAverageTest() {
         SalesCalculate service = new SalesCalculate();
         long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
-        long expected = 5;
-        long actual = service.monthsSalesOverAverage(sales);
+        int expected = 5;
+        int actual = service.monthsSalesOverAverage(sales);
         Assertions.assertEquals(expected, actual);
     }
 
@@ -54,8 +54,8 @@ public class SalesCalculateTest {
     void monthsSalesUnderAverageTest() {
         SalesCalculate service = new SalesCalculate();
         long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
-        long expected = 5;
-        long actual = service.monthsSalesUnderAverage(sales);
+        int expected = 5;
+        int actual = service.monthsSalesUnderAverage(sales);
         Assertions.assertEquals(expected, actual);
     }
 }
